Create README.md
Weather Forecast is a web application that provides up-to-date weather information for a selected city. Users can enter the name of a city in the input form, and the application will send a request to a weather API to fetch and display the current temperature, feels like temperature, cloud conditions, and humidity. The application is built using HTML, CSS, and JavaScript and utilizes a weather API to retrieve weather data. The design was also developed by me.